% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbr2bib.R
\name{abbr2bib}
\alias{abbr2bib}
\title{The BibTeX \code{tibble} is written to a .bib file}
\usage{
abbr2bib(file, outfile_abbr = "abbr.bib", separate_names = FALSE)
}
\arguments{
\item{file}{character, path or URL to a .bib file.}

\item{outfile_abbr}{character, file path to write the .bib file. An empty character string writes to \code{stdout} (default).}

\item{separate_names}{logical, should authors' and editors' names be separated into first and given name?}
}
\value{
{
Return to a list and  output the BibTex file after the abbreviation of Journal.
List consists of three data structures as follows:
}
\describe{
\item{abbrtable}{A data.table, the abbreviation table of journal}
\item{noabbr}{A data.table, No abbreviated journals found}
\item{noindex}{A vector, Index corresponding to noabbr}
}
}
\description{
Input Bib file with complete journal, output Bib file after abbreviation of journa, and return to the abbreviation table of journal
}
\examples{
\dontrun{
library(journalabbr)
path = system.file("extdata", "testfile_1.bib", package = "journalabbr", mustWork = TRUE)
abbr2bib(file = path,outfile_abbr= "abbr.bib")
}
}
\keyword{BibTex}
\keyword{List}
\keyword{abbreviations}
\keyword{file}
\keyword{for}
\keyword{input}
\keyword{journal}
\keyword{of}
