% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abbrTable.R
\docType{data}
\name{abbrTable}
\alias{abbrTable}
\title{data abbrTable data}
\format{
A data frame with columns:
\describe{
\item{journal}{Full name of Journal}
\item{journal_abbr}{Abbreviated format of Journal with dots}
\item{journal_lower}{The full name of the journal should be in lowercase and unique}
\item{count_dot}{The number of points in abbreviations of Journal with dots}
\item{originFile}{File source}
\item{abbr_len}{The char length  of abbreviations of Journal }
}
}
\source{
https://github.com/JabRef/abbrv.jabref.org/journals
}
\usage{
abbrTable
}
\description{
There are four columns, each separated by a semicolon (;), as follows
journal  journal_abbr journal_lower. count_dot
}
\keyword{datasets}
