% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_abbrtable.R
\name{add_abbrtable}
\alias{add_abbrtable}
\title{Add user's journal abbreviation table}
\usage{
add_abbrtable(file, ...)
}
\arguments{
\item{file}{character, bib file path}

\item{...}{see also \code{data.table::fread}}
}
\value{
a data.table
}
\description{
Add user's journal abbreviation table
}
\examples{
csvpath <- system.file("extdata", "myabbr.csv", package = "journalabbr", mustWork = TRUE)
abbrtable_user <- add_abbrtable(file = csvpath, header = FALSE, sep = ";")
colnames(abbrtable_user)

}
