% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites_convert.R
\name{single_cites_convert}
\alias{single_cites_convert}
\alias{cites_convert}
\title{Convert citations between LaTeX and Markdown formats}
\usage{
single_cites_convert(
  text,
  pattern = NULL,
  tex2md = TRUE,
  latex_prefix = c("cite", "upcite", "citep", "citet")
)

cites_convert(x, ...)
}
\arguments{
\item{text}{A character string containing the text with citations, Note that the input text must have a length of 1.}

\item{pattern}{A regular expression pattern for matching citations. If NULL, defaults to a pattern based on the citation direction.}

\item{tex2md}{Logical; if TRUE, converts from LaTeX to Markdown; if FALSE, converts from Markdown to LaTeX.}

\item{latex_prefix}{A character vector of LaTeX citation prefixes to use.
Defaults to c("cite", "upcite", "citep", "citet").}

\item{x}{A character vector containing texts with citations.}

\item{...}{Additional arguments passed to \code{single_cites_convert}.}
}
\value{
the \code{single_cites_convert} function return a character string with citations converted to the specified format.

the \code{cites_convert} function  return a character vector with citations converted to the specified format.
}
\description{
This function converts citations in a given text between LaTeX and Markdown formats.
It can handle multiple LaTeX citation styles and produces the appropriate format based on the specified direction.
Note that the input text must have a length of 1.

the \code{cites_convert} function applies \code{single_cites_convert} to each element of the input vector.
}
\examples{
# Example text containing various LaTeX citations
# Convert the LaTeX citations in the text to Markdown format
text = "This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }.
        paper \\\\cite{key3-b, cky3-c}, paper \\\\citet{key4-a, cky4-b}"
converted_md = single_cites_convert(text, tex2md = TRUE)
print(converted_md)



# Example text containing Markdown citations
# Convert the Markdown citations back to LaTeX format

text = "This is a citation [@key1; @key2] and [@key3-a].
        paper [@key3-b; @cky3-c], paper [@key4-a; @cky4-b]"
# defalut prefix "cite"
converted_latex = single_cites_convert(text, tex2md = FALSE)
print(converted_latex)

# Convert Markdown citations to LaTeX using a custom prefix
custom_prefix = single_cites_convert(text, tex2md = FALSE, latex_prefix = "upcite")
print(custom_prefix)


# A vector of texts with Markdown citations
tex =c("This is a citation [@key1; @key2] and [@key3-a].",
        "This is a citation2 [@key1; @key2] and [@key3-a].",
        "this is no citation",
        "",
        "This is a citation3 [@key1; @key2] and [@key3-a].")

# vector version
tex = cites_convert(tex, tex2md = FALSE)
print(tex)



# Another vector of texts containing LaTeX citations
tex2= c("This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }",
         "This is a citation \\\\citet{key1,key2} and \\\\citet{key3-a }",
         "this is no citation",
         "",
         "This is a citation \\\\upcite{key1,key2} and \\\\cite{key3-a }")
# Convert the vector of LaTeX citations to Markdown format
tex2 = cites_convert(tex2, tex2md = TRUE)
print(tex2)


}
