% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites_extract.R
\name{single_cites_extract}
\alias{single_cites_extract}
\alias{cites_extract}
\title{Extract Single Citation Keys from LaTeX Text}
\usage{
single_cites_extract(text)

cites_extract(tex)
}
\arguments{
\item{text}{A character vector containing LaTeX content.}

\item{tex}{A character vector where each element represents a line or section of LaTeX content.}
}
\value{
A character vector of citation keys, or \code{NULL} if no citations are found.

A unique character vector of citation keys.
}
\description{
\code{single_cites_extract} function extracts all citation keys from a LaTeX string based on specific citation commands.

\code{cites_extract} function extracts all unique citation keys from a LaTeX document.
}
\details{
The function identifies citation commands such as \code{cite}, \code{upcite}, \code{citep}, and \code{citet}.
It extracts the content within braces following these commands, and splits multiple keys separated by commas.

This function uses \code{single_cites_extract} to extract citation keys from each line
or section of the LaTeX document. It removes duplicates and trims unnecessary whitespace.
}
\examples{
tex= c("This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }",
      "This is a citation \\\\citet{key1,key2} and \\\\citet{key3-a }",
      "this is no citation",
      "",
      "This is a citation \\\\upcite{key1,key4} and \\\\cite{key3-a }")
cites_extract(tex)


tex= c("This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }",
      "This is a citation \\\\citet{key1,key2} and \\\\citet{key3-a }",
      "this is no citation",
      "",
      "This is a citation \\\\upcite{key1,key4} and \\\\cite{key3-a }")
cites_extract(tex)


}
