% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cites_replace_order.R
\name{cites_replace_order}
\alias{cites_replace_order}
\title{Replace All Citations in a LaTeX Document Based on Order}
\usage{
cites_replace_order(tex, citesall = NULL)
}
\arguments{
\item{tex}{A character vector where each element represents a line or section of LaTeX content.}

\item{citesall}{A character vector containing all citation keys in the desired order. If not provided, default applies \code{cites_extract}.}
}
\value{
A character vector with updated LaTeX content where citation keys in commands are reordered.
}
\description{
\code{cites_replace_order} function processes a LaTeX document to reorder all citation keys within citation commands
based on their first appearance in the document.
}
\details{
The function extracts all citation keys using \code{cites_extract}, determines their order based on first occurrence,
and applies \code{single_replace_order} to reorder keys in each citation command.
}
\examples{
tex= c("This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }",
       "This is a citation \\\\citet{key1,key2} and \\\\citet{key3-a }",
       "this is no citation",
       "",
       "This is a citation \\\\upcite{key1,key4} and \\\\cite{key3-a }")
citesall = cites_extract(tex)

citesall = c("key2", "key1", "key3-a", "key4")

cites_replace_order(tex, citesall)
}
\seealso{
\code{\link{single_replace_order}}, \code{\link{cites_extract}}
}
