% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_delcolumns.R
\name{get_delcolumns}
\alias{get_delcolumns}
\title{Get default columns to delete}
\usage{
get_delcolumns()
}
\value{
A character vector containing the default column names to be excluded.
}
\description{
Returns the default list of columns that will be deleted.
}
