% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journal_abbreviations_lower_all.R
\docType{data}
\name{journal_abbreviations_lower_all}
\alias{journal_abbreviations_lower_all}
\title{data journal_abbreviations_lower_all data}
\format{
A data frame with columns:
\describe{
\item{journal}{Full name of Journal}
\item{journal_abbr}{Abbreviated format of Journal with dots}
\item{journal_lower}{The full name of the journal should be in lowercase and unique}
\item{count_dot}{The number of points in abbreviations of Journal with dots}
}
}
\source{
https://github.com/JabRef/abbrv.jabref.org/journals
}
\usage{
journal_abbreviations_lower_all
}
\description{
There are four columns, each separated by a semicolon (;), as follows
journal  journal_abbr journal_lower. count_dot
}
\keyword{datasets}
