% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/journalabbr-package.R
\docType{package}
\name{journalabbr-package}
\alias{journalabbr}
\alias{journalabbr-package}
\title{journalabbr: Journal Abbreviations for BibTeX Documents}
\description{
Since the reference management software (such as 'Zotero', 'Mendeley') exports Bib file journal abbreviation is not detailed enough, the 'journalabbr' package only abbreviates the journal field of Bib file, and then outputs a new Bib file for generating reference format with journal abbreviation on other software (such as 'texstudio'). The abbreviation table is from 'JabRef'. At the same time, 'Shiny' application is provided to generate 'thebibliography', a reference format that can be directly used for latex paper writing based on 'Rmd' files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zoushucai/journalabbr}
  \item Report bugs at \url{https://github.com/zoushucai/journalabbr/issues}
}

}
\author{
\strong{Maintainer}: ShuCai Zou \email{zscmoyujian@163.com}

Authors:
\itemize{
  \item Yu Chen \email{chengyuxuezhe@163.com}
}

}
\keyword{internal}
