% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bib2tib.R
\name{read_bib2tib}
\alias{read_bib2tib}
\title{Parse a BibTeX file to a \code{tibble}}
\usage{
read_bib2tib(file, isabbr = FALSE)
}
\arguments{
\item{file}{character, path or URL to a .bib file.}

\item{isabbr}{logical,  if \code{TRUE}, If true, two columns \code{journal_abbr} and \code{originFile} will be added. \code{journal_abbr} is the abbreviation for the journal and \code{originFile} is the source of the abbreviated journal}
}
\value{
A \code{tibble}.
}
\description{
The BibTeX file is read, parsed, tidied and written to a \code{tibble}
}
\details{
For simplicity \code{read_bib2tib()} unifies the reading, parsing and tidying of a BibTeX file while being aware of a standardized output format, different BibTeX styles and missing values in the BibTeX file.
}
\examples{
# Read from .bib file:
require(journalabbr)
path = system.file("extdata", "testfile_1.bib", package = "journalabbr", mustWork = TRUE)
bib <- read_bib2tib(path)
str(bib)
}
\seealso{
\code{\link{abbrTable}}
}
\author{
ShuCai Zou
}
