% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex2md.R
\name{single_convert_citations}
\alias{single_convert_citations}
\title{Convert citations between LaTeX and Markdown formats}
\usage{
single_convert_citations(
  text,
  pattern = NULL,
  tex2md = TRUE,
  latex_prefix = c("cite", "upcite", "citep", "citet")
)
}
\arguments{
\item{text}{A character string containing the text with citations, Note that the input text must have a length of 1.}

\item{pattern}{A regular expression pattern for matching citations. If NULL, defaults to a pattern based on the citation direction.}

\item{tex2md}{Logical; if TRUE, converts from LaTeX to Markdown; if FALSE, converts from Markdown to LaTeX.}

\item{latex_prefix}{A character vector of LaTeX citation prefixes to use.
Defaults to c("cite", "upcite", "citep", "citet").}
}
\value{
the \code{convert_citations_single} function return a character string with citations converted to the specified format.
}
\description{
This function converts citations in a given text between LaTeX and Markdown formats.
It can handle multiple LaTeX citation styles and produces the appropriate format based on the specified direction.
Note that the input text must have a length of 1.
}
\examples{
# Example text containing various LaTeX citations
# Convert the LaTeX citations in the text to Markdown format
text = "This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }.
        paper \\\\cite{key3-b, cky3-c}, paper \\\\citet{key4-a, cky4-b}"
converted_md = convert_citations(text, tex2md = TRUE)
print(converted_md)



# Example text containing Markdown citations
# Convert the Markdown citations back to LaTeX format

text = "This is a citation [@key1; @key2] and [@key3-a].
        paper [@key3-b; @cky3-c], paper [@key4-a; @cky4-b]"
# defalut prefix "cite"
converted_latex = convert_citations(text, tex2md = FALSE)
print(converted_latex)

# Convert Markdown citations to LaTeX using a custom prefix
custom_prefix = convert_citations(text, tex2md = FALSE, latex_prefix = "upcite")
print(custom_prefix)


# A vector of texts with Markdown citations
text_vector =c("This is a citation [@key1; @key2] and [@key3-a].",
        "This is a citation2 [@key1; @key2] and [@key3-a].",
        "this is no citation",
        "",
        "This is a citation3 [@key1; @key2] and [@key3-a].")

# vector version
vector_latex = convert_citations_vec(text_vector, tex2md = FALSE)
print(vector_latex)



# Another vector of texts containing LaTeX citations
text_vector2= c("This is a citation \\\\upcite{key1,key2} and \\\\citep{key3-a }",
         "This is a citation \\\\citet{key1,key2} and \\\\citet{key3-a }",
         "this is no citation",
         "",
         "This is a citation \\\\upcite{key1,key2} and \\\\cite{key3-a }")
# Convert the vector of LaTeX citations to Markdown format
vector_md = convert_citations_vec(text_vector2, tex2md = TRUE)
print(vector_md)



texfile = "C:\\\\Users\\\\zscmm\\\\Documents\\\\GitHub\\\\journalabbr\\\\weak.tex"
tex = readLines(texfile)
newtex = cites_order_replace(tex)
texfile2 = "C:\\\\Users\\\\zscmm\\\\Documents\\\\GitHub\\\\journalabbr\\\\weak2.tex"
writeLines(newtex, con = texfile2)


texfile = "C:\\\\Users\\\\zscmm\\\\Documents\\\\GitHub\\\\journalabbr\\\\weak2-author.tex"
tex = readLines(texfile)
text = tex[157]
print(text)

newtex = cite_replace_author(tex)
texfile2 = "C:\\\\Users\\\\zscmm\\\\Documents\\\\GitHub\\\\journalabbr\\\\weak2-back.tex"
writeLines(newtex, con = texfile2)
}
