% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex_funs.R
\name{single_add_space}
\alias{single_add_space}
\alias{tex_funs}
\title{Clean up and standardize spacing in a given text}
\usage{
single_add_space(text)

tex_funs(tex, fun = single_add_space, ...)
}
\arguments{
\item{text}{A single character string (input text) that needs to be processed.}

\item{tex}{A vector of strings (LaTeX document or text) to be processed.}

\item{fun}{The function to be applied to each string in the vector (default is \code{single_add_space}).}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A cleaned-up string where multiple spaces are reduced to one, unnecessary spaces
are removed, and spaces are appropriately added around certain punctuation and
mathematical symbols.

A vector of processed strings where unnecessary spaces have been removed or adjusted.
}
\description{
This function processes a given string by cleaning up extra spaces, handling punctuation,
and adjusting mathematical symbols according to specific formatting rules.

This function takes a vector of LaTeX-formatted strings and applies the \code{single_add_space}
function to each string to standardize spacing throughout the document.
}
\details{
see function \code{single_add_space} source code for details.
}
\examples{
tex =c("This is a citation,and.",
       "This is a citation2 $x ^ 2+y^ 2 = z^2.$ and this is equ",
       "this is no , citation")
tex = tex_funs(tex)
print(tex)
}
